requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.13.1
google-cloud-run==0.10.0
pinecone-client==2.2.4
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1